requirements.txt
# AI Portfolio Manager - Requirements

# Core Data Processing
numpy>=1.24.0,<2.0.0
pandas>=2.0.0,<2.3.0
scipy>=1.10.0,<2.0.0

# Data Acquisition
yfinance>=0.2.28
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Technical Analysis (TA-Lib is optional, system works without it)
# TA-Lib requires: sudo apt-get install ta-lib (Ubuntu) or brew install ta-lib (macOS)
# Uncomment if you have TA-Lib installed:
# TA-Lib>=0.4.28

# Machine Learning - Core
scikit-learn>=1.3.0,<1.6.0
joblib>=1.3.0

# Deep Learning - PyTorch (for LSTM/GRU)
torch>=2.0.0
torchvision>=0.15.0

# Gradient Boosting
xgboost>=2.0.0
lightgbm>=4.0.0
catboost>=1.2.0

# NLP & Sentiment Analysis
transformers>=4.30.0
sentencepiece>=0.1.99
vaderSentiment>=3.3.2

# Portfolio Optimization
PyPortfolioOpt>=1.5.5
cvxpy>=1.4.0

# Reinforcement Learning
stable-baselines3>=2.0.0
gymnasium>=0.28.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Configuration & Utils
pyyaml>=6.0.0
python-dotenv>=1.0.0
tqdm>=4.65.0

# Performance (optional)
numba>=0.57.0
